refactor(indexer): rename indexer_reader to read
