+++ crates/iota-indexer/src/indexer.rs
@@ -24,9 +24,9 @@ use crate::{
         checkpoint_handler::new_handlers, objects_snapshot_handler::start_objects_snapshot_handler,
         optimistic_pruner::OptimisticPruner, pruner::Pruner,
     },
-    indexer_reader::IndexerReader,
     metrics::IndexerMetrics,
     processors::processor_orchestrator::ProcessorOrchestrator,
+    read::IndexerReader,
     store::{IndexerAnalyticalStore, IndexerStore, PgIndexerStore},
 };
 
@@ -156,8 +156,8 @@ impl Indexer {
             "IOTA Indexer Reader (version {:?}) started...",
             env!("CARGO_PKG_VERSION")
         );
-        let indexer_reader = IndexerReader::new(connection_pool);
-        let handle = build_json_rpc_server(store, registry, indexer_reader, config, metrics)
+        let read = IndexerReader::new(connection_pool);
+        let handle = build_json_rpc_server(store, registry, read, config, metrics)
             .await
             .expect("Json rpc server should not run into errors upon start.");
         tokio::spawn(async move { handle.stopped().await })
